consolidate upload util files
